Breaking down sync into multiple structs and files
